task/1: chore - set up minimal code for web and api
