+++ README.md
@@ -1,39 +1,21 @@
-# Turborepo kitchen sink starter
+Hello there 👋 <br/>
+Welcome to my Scraper code repo. <br/>
 
-This is an official starter Turborepo with multiple meta-frameworks all working in harmony and sharing packages.
+Scrapper is a mini web platform that let you upload a CSV file with the keywords inside, and use that CSV file to scrape on Google to give analytics on search result.
 
-This example also shows how to use [Workspace Configurations](https://turbo.build/repo/docs/core-concepts/monorepos/configuring-workspaces).
-
-## Using this example
-
-Run the following command:
-
-```sh
-npx create-turbo@latest -e kitchen-sink
+Here is a sample format of the CSV file (each keyword is line-separated):
+```csv
+macbook air m2
+laptop
+iphone
+iphone 15 pink
 ```
 
-## What's inside?
+If you are curious about my thought process or what I have been working for this project, here are some resources for you to get started:
+- [Wiki](https://github.com/nhantran3395/scrapper/wiki)
+- [Scrum Board](https://github.com/users/nhantran3395/projects/2)
+- [How to get started?](https://github.com/nhantran3395/scrapper/wiki/%5BGuide%5D-How-to-get-started%3F)
 
-This Turborepo includes the following packages and apps:
+Check out the [How to get started?](https://github.com/nhantran3395/scrapper/wiki/%5BGuide%5D-How-to-get-started%3F) page on guide to set up and run Scrapper on your local
 
-### Apps and Packages
-
-- `api`: an [Express](https://expressjs.com/) server
-- `storefront`: a [Next.js](https://nextjs.org/) app
-- `admin`: a [Vite](https://vitejs.dev/) single page app
-- `blog`: a [Remix](https://remix.run/) blog
-- `logger`: isomorphic logger (a small wrapper around console.log)
-- `ui`: a dummy React UI library (which contains a single `<CounterButton>` component)
-- `scripts`: Jest and ESLint configurations
-- `tsconfig`: tsconfig.json's used throughout the monorepo
-
-Each package and app is 100% [TypeScript](https://www.typescriptlang.org/).
-
-### Utilities
-
-This Turborepo has some additional tools already setup for you:
-
-- [TypeScript](https://www.typescriptlang.org/) for static type checking
-- [ESLint](https://eslint.org/) for code linting
-- [Jest](https://jestjs.io) test runner for all things JavaScript
-- [Prettier](https://prettier.io) for code formatting
+Check out the [The main challenges and approaches](https://github.com/nhantran3395/scrapper/wiki/%5BDesign%5D-The-main-challenges-and-approaches) page to find out how I design and solve for the technical challenges
