--- apps/admin/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "admin",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "build": "vite build",
-    "clean": "rm -rf dist",
-    "dev": "vite --host 0.0.0.0 --port 3001 --clearScreen false",
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint src/"
-  },
-  "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "ui": "*"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.21",
-    "@types/react-dom": "^18.2.7",
-    "@vitejs/plugin-react": "^4.0.4",
-    "eslint-config-custom": "*",
-    "tsconfig": "*",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.9"
-  }
-}
