--- apps/admin/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "exclude": ["node_modules"],
-  "extends": "tsconfig/vite.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src"]
-}
