--- apps/admin/vite.config.ts
@@ -1,6 +1,0 @@
-import react from "@vitejs/plugin-react";
-import { defineConfig } from "vite";
-
-export default defineConfig({
-  plugins: [react()],
-});
