+++ apps/api/src/__tests__/server.test.ts
@@ -10,13 +10,4 @@ describe("Server", () => {
         expect(res.ok).toBe(true);
       });
   });
-
-  it("message endpoint says hello", async () => {
-    await supertest(createServer())
-      .get("/message/jared")
-      .expect(200)
-      .then((res) => {
-        expect(res.body).toEqual({ message: "hello jared" });
-      });
-  });
 });
