--- apps/blog/.gitignore
@@ -1,11 +1,0 @@
-node_modules
-
-.cache
-.env
-.vercel
-.output
-
-build
-public/build
-api/index.js*
-api/_assets*
