--- apps/blog/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "blog",
-  "version": "0.0.0",
-  "private": true,
-  "sideEffects": false,
-  "scripts": {
-    "build": "remix build",
-    "dev": "remix dev",
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint app/"
-  },
-  "dependencies": {
-    "@remix-run/node": "^1.19.3",
-    "@remix-run/react": "^1.19.3",
-    "@remix-run/serve": "^1.19.3",
-    "@remix-run/server-runtime": "^1.19.3",
-    "@vercel/analytics": "^0.1.11",
-    "@vercel/remix-entry-server": "^0.1.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "ui": "*"
-  },
-  "devDependencies": {
-    "@remix-run/dev": "^1.19.3",
-    "@types/react": "^18.2.21",
-    "@types/react-dom": "^18.2.7",
-    "eslint-config-custom": "*",
-    "tsconfig": "*",
-    "typescript": "^5.2.2"
-  },
-  "engines": {
-    "node": ">=18"
-  }
-}
