--- packages/tsconfig/vite.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "./base.json",
-  "Display": "Vite",
-  "compilerOptions": {
-    "allowJs": false,
-    "esModuleInterop": false,
-    "jsx": "react",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "module": "ESNext",
-    "noEmit": true,
-    "resolveJsonModule": true,
-    "skipLibCheck": false,
-    "target": "ESNext",
-    "types": ["vite/client"]
-  }
-}
