--- packages/ui/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "ui",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "sideEffects": false,
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist/**"
-  ],
-  "scripts": {
-    "build": "tsup src/index.tsx --format esm,cjs --dts --external react",
-    "clean": "rm -rf dist",
-    "dev": "tsup src/index.tsx --format esm,cjs --watch --dts --external react",
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint src/",
-    "test": "jest"
-  },
-  "jest": {
-    "preset": "jest-presets/jest/browser"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.21",
-    "@types/react-dom": "^18.2.7",
-    "eslint-config-custom": "*",
-    "jest-presets": "*",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tsconfig": "*",
-    "tsup": "^6.2.3",
-    "typescript": "^5.2.2"
-  }
-}
