--- packages/ui/src/CounterButton/index.test.tsx
@@ -1,11 +1,0 @@
-import { createRoot } from "react-dom/client";
-import { CounterButton } from ".";
-
-describe("CounterButton", () => {
-  it("renders without crashing", () => {
-    const div = document.createElement("div");
-    const root = createRoot(div);
-    root.render(<CounterButton />);
-    root.unmount();
-  });
-});
