--- packages/ui/src/index.tsx
@@ -1,2 +1,0 @@
-export { Link } from "./Link";
-export { CounterButton } from "./CounterButton";
