--- packages/ui/tsup.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig, type Options } from "tsup";
-
-export default defineConfig((options: Options) => ({
-  banner: {
-    js: "'use client'",
-  },
-  ...options,
-}));
