--- packages/ui/turbo.json
@@ -1,8 +1,0 @@
-{
-  "extends": ["//"],
-  "pipeline": {
-    "build": {
-      "outputs": ["dist/**"]
-    }
-  }
-}
